requirements.txt
tensorflow~=2.18.0
numpy~=2.0.2
streamlit~=1.41.1
torch~=2.5.1
pillow~=11.1.0
tqdm~=4.67.1
transformers~=4.47.1